requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.4.2
Sphinx==1.6.5
twine==1.9.1
cryptography==2.1.4
PyYAML==3.11
Adafruit-BBIO==1.0.10
google-cloud-datastore==1.4.0
google-cloud-storage==1.6.0
numpy==1.13.3
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
PyBOMBS==2.3.2
pyserial==3.4
python-apt==1.4.0b3

